+++ packages/config/src/xcm-configs/astar.ts
@@ -66,7 +66,10 @@ export const astarRoutes = new ChainRoutes({
           asset: glmr,
         },
       },
-      extrinsic: ExtrinsicBuilder().xTokens().transfer(),
+      extrinsic: ExtrinsicBuilder()
+        .polkadotXcm()
+        .limitedReserveTransferAssets()
+        .X2PalletInstance(),
     },
   ],
 });