+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @moonbeam-network/xcm-sdk
 
+## 3.2.8
+
+### Patch Changes
+
+- Updated dependencies [[`8179c3f`](https://github.com/moonbeam-foundation/xcm-sdk/commit/8179c3f97be3e5d9c23db1cce1d868ea013f15e3)]:
+  - @moonbeam-network/xcm-builder@3.3.6
+  - @moonbeam-network/xcm-config@3.4.2
+  - @moonbeam-network/xcm-types@3.2.6
+
 ## 3.2.7
 
 ### Patch Changes